tackle comments

+++ test/mp_unit/proxy_channel_tests.cu
@@ -211,7 +211,7 @@ __global__ void kernelProxyLLPingPong(int* buff, mscclpp::LLPacket* putPktBuf, m
   int threadId = threadIdx.x + blockIdx.x * blockDim.x;
   int numThreads = blockDim.x * gridDim.x;
   int flusher = 0;
-  const size_t nPkt = nElem / 2;
+  const int nPkt = nElem / 2;
   for (int i = 0; i < nTries; i++) {
     uint64_t flag = (uint64_t)i + 1;
 
@@ -220,9 +220,9 @@ __global__ void kernelProxyLLPingPong(int* buff, mscclpp::LLPacket* putPktBuf, m
     if ((rank ^ (i & 1)) == 0) {
       if (CheckCorrectness) {
         // If each thread writes 8 bytes at once, we don't need a barrier before putPackets().
-        for (size_t j = threadId; j < nPkt; j += numThreads) {
-          buffPtr[2 * j] = putOffset + i + 2 * (int)j;
-          buffPtr[2 * j + 1] = putOffset + i + 2 * (int)j + 1;
+        for (int j = threadId; j < nPkt; j += numThreads) {
+          buffPtr[2 * j] = putOffset + i + 2 * j;
+          buffPtr[2 * j + 1] = putOffset + i + 2 * j + 1;
         }
         // __syncthreads();
       }
@@ -242,14 +242,14 @@ __global__ void kernelProxyLLPingPong(int* buff, mscclpp::LLPacket* putPktBuf, m
       if (CheckCorrectness) {
         // If each thread reads 8 bytes at once, we don't need a barrier after getPackets().
         // __syncthreads();
-        for (size_t j = threadId; j < nPkt; j += numThreads) {
-          if (buffPtr[2 * j] != getOffset + i + 2 * (int)j) {
+        for (int j = threadId; j < nPkt; j += numThreads) {
+          if (buffPtr[2 * j] != getOffset + i + 2 * j) {
             // printf("ERROR: rank = %d, buffPtr[%d] = %d, expected %d. Skipping following errors\n", rank, 2 * j,
             //        buffPtr[2 * j], getOffset + i + 2 * j);
             *ret = 1;
             break;
           }
-          if (buffPtr[2 * j + 1] != getOffset + i + 2 * (int)j + 1) {
+          if (buffPtr[2 * j + 1] != getOffset + i + 2 * j + 1) {
             // printf("ERROR: rank = %d, buffPtr[%d] = %d, expected %d. Skipping following errors\n", rank, 2 * j + 1,
             //        buffPtr[2 * j + 1], getOffset + i + 2 * j + 1);
             *ret = 1;
